minor fix for xqa (#1994)
<!-- .github/pull_request_template.md -->

## 📌 Description

<!-- What does this PR do? Briefly describe the changes and why they’re
needed. -->
1 change xqa_mla comments to be consistent with mla instead of mha.
2 put cudaMemcpyFromSymbol/cudaFuncSetAttribute outside of launch
function to avoid breaking cuda graph capture
3 use int32 as pagetable index 

## 🔍 Related Issues

<!-- Link any related issues here -->

## 🚀 Pull Request Checklist

Thank you for contributing to FlashInfer! Before we review your pull
request, please make sure the following items are complete.

### ✅ Pre-commit Checks

- [x] I have installed `pre-commit` by running `pip install pre-commit`
(or used your preferred method).
- [x] I have installed the hooks with `pre-commit install`.
- [x] I have run the hooks manually with `pre-commit run --all-files`
and fixed any reported issues.

> If you are unsure about how to set up `pre-commit`, see [the
pre-commit documentation](https://pre-commit.com/).

## 🧪 Tests

- [x] Tests have been added or updated as needed.
- [ ] All tests are passing (`unittest`, etc.).

## Reviewer Notes

<!-- Optional: anything you'd like reviewers to focus on, concerns, etc.
-->

+++ tests/attention/test_xqa.py
@@ -253,7 +253,7 @@ def test_xqa(
         cache_k_heads /= 4.0
         cache_v_heads /= 4.0
     page_list_arg = torch.zeros(
-        batch_size, nb_pages_per_seq, dtype=torch.uint32, device="cuda"
+        batch_size, nb_pages_per_seq, dtype=torch.int32, device="cuda"
     )
 
     # Initialize page list sequentially
@@ -265,7 +265,7 @@ def test_xqa(
 
     flattened = page_list_arg.flatten()
     indices = torch.randperm(flattened.numel())
-    shuffled_flat = flattened.to(torch.int32)[indices].to(torch.uint32)
+    shuffled_flat = flattened[indices]
     page_list_arg = shuffled_flat.view(page_list_arg.shape)
 
     def cache_head_at(
@@ -470,7 +470,7 @@ def test_xqa_mla(
     cache_v_heads /= 4.0
 
     page_list_arg = torch.zeros(
-        batch_size, nb_pages_per_seq, dtype=torch.uint32, device="cuda"
+        batch_size, nb_pages_per_seq, dtype=torch.int32, device="cuda"
     )
 
     # Initialize page list sequentially
@@ -482,7 +482,7 @@ def test_xqa_mla(
 
     flattened = page_list_arg.flatten()
     indices = torch.randperm(flattened.numel())
-    shuffled_flat = flattened.to(torch.int32)[indices].to(torch.uint32)
+    shuffled_flat = flattened[indices]
     page_list_arg = shuffled_flat.view(page_list_arg.shape)
 
     def cache_head_at(
